+++ CHANGELOG.md
@@ -6,6 +6,14 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Added
+ - Support for [Authenticated Users](https://docs.pusher.com/beams/concepts/authenticated-users):
+ `generateToken`, `publishToUsers` and `deleteUser` functions were added
+
+### Changed
+ - Renamed `publish` method to `publishToInterest`
+
+
 ## [1.0.0] - 2018-07-31
 ### Added
  - Changelog for GA release